requirements.txt
kabbes_dir_ops>=0.7.1
kabbes_parent_class>=0.6.0
kabbes_py_starter>=0.10.0
